fetch german names s
