+++ backend/src/model/enum/propagation_method.rs
@@ -1,3 +1,5 @@
+//! [`PropagationMethod`] enum.
+
 use diesel_derive_enum::DbEnum;
 use serde::{Deserialize, Serialize};
 use typeshare::typeshare;
