+++ scraper/src/fetch_practicalplants.js
@@ -2,41 +2,11 @@ import { readFileSync, readdirSync, writeFileSync, existsSync, mkdirSync } from 
 import { parse } from 'node-html-parser';
 import { parse as json2csv } from 'json2csv';
 import { config } from 'dotenv';
-import axios from 'axios';
+import { fetchGermanName } from './helpers/helpers.js';
 
 config();
 
 const archivePath = process.env.PRACTICALPLANTSPATH;
-
-/**
- * Fetches the German name of the plant from wikidata API.
- *
- * @param {*} binomialName
- * @returns {Promise<string>} German name of the plant
- */
-async function fetchGermanName(binomialName) {
-  try {
-    const url = `https://www.wikidata.org/w/api.php?action=wbsearchentities&search=${binomialName}&language=en&format=json`;
-    const response = await axios.get(url);
-    const data = response.data;
-    const results = data.search;
-    if (results.length === 0) {
-      return null;
-    }
-    const result = results[0];
-    const id = result.id;
-    const url2 = `https://www.wikidata.org/w/api.php?action=wbgetentities&ids=${id}&languages=de&format=json`;
-    const response2 = await axios.get(url2);
-    const data2 = response2.data;
-    const entities = data2.entities;
-    const entity = entities[id];
-    const dewiki = await entity['sitelinks']['dewiki'];
-    if (dewiki) {
-      return dewiki.title;
-    }
-  } catch (error) {}
-  return null;
-}
 
 function processData(details) {
   try {
