+++ scraper/src/helpers/helpers.js
@@ -1,3 +1,5 @@
+import axios from 'axios';
+
 /**
  * Sanitizes the column names of the json array
  *
@@ -72,4 +74,34 @@ function getSoilPH(pH) {
   }
 }
 
-export { sanitizeColumnNames, getSoilPH };
+/**
+ * Fetches the German name of the plant from wikidata API.
+ *
+ * @param {*} binomialName
+ * @returns {Promise<string>} German name of the plant
+ */
+async function fetchGermanName(binomialName) {
+  try {
+    const url = `https://www.wikidata.org/w/api.php?action=wbsearchentities&search=${binomialName}&language=en&format=json`;
+    const response = await axios.get(url);
+    const data = response.data;
+    const results = data.search;
+    if (results.length === 0) {
+      return null;
+    }
+    const result = results[0];
+    const id = result.id;
+    const url2 = `https://www.wikidata.org/w/api.php?action=wbgetentities&ids=${id}&languages=de&format=json`;
+    const response2 = await axios.get(url2);
+    const data2 = response2.data;
+    const entities = data2.entities;
+    const entity = entities[id];
+    const dewiki = await entity['sitelinks']['dewiki'];
+    if (dewiki) {
+      return dewiki.title;
+    }
+  } catch (error) {}
+  return null;
+}
+
+export { sanitizeColumnNames, getSoilPH, fetchGermanName };
